% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{getDiffRhythmStats}
\alias{getDiffRhythmStats}
\title{Calculate differential rhythmic statistics between conditions}
\usage{
getDiffRhythmStats(fit, rhyStats, condLevels)
}
\arguments{
\item{fit}{A limorhyde2 fit object, as provided by \code{getModelFit} or
\code{getPosteriorFit}.}

\item{rhyStats}{A data.table of rhythmic statistics, as returned by
\code{getRhythmStats}. The \code{fitType} of rhyStats is inherited by the returned
object.}

\item{condLevels}{A character vector containing the two conditions to compare.}
}
\value{
A data.table of differential rhythmic statistics (differential mean value,
differential peak-trough amplitude, differential root mean-squared amplitude,
differential peak phase, and differential trough phase) between paired conditions.
}
\description{
\code{getDiffRhythmStats} returns the differential rhythmic statistics
calculated under differing conditions.
}
