% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitted.R
\name{getFittedIntervals}
\alias{getFittedIntervals}
\title{Get credible intervals for fitted values}
\usage{
getFittedIntervals(fittedVals, mass = 0.9, method = c("eti", "hdi"))
}
\arguments{
\item{fittedVals}{A data.table of posterior samples for fitted values from
\code{getFittedValues()}.}

\item{mass}{The probability mass for which to calculate the interval.}

\item{method}{"String for type of interval: 'eti' for equal-tailed or 'hdi' for
highest (posterior) density.}
}
\value{
A data.table with columns for the upper and lower bounds of the fitted
value for each feature in each group.
}
\description{
\code{getFittedIntervals()} constructs credible intervals for the fitted values of
a model by group.
}
\seealso{
\code{\link[=getFittedValues]{getFittedValues()}}
}
