% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitted.R
\name{getFittedValues}
\alias{getFittedValues}
\title{Compute the predicted values from a fitted model}
\usage{
getFittedValues(
  fit,
  times,
  fitType = c("posterior_mean", "posterior_samples", "raw"),
  features = NULL
)
}
\arguments{
\item{fit}{A 'limorhyde2' object, as provided by \code{getModelFit()} or
\code{getPosteriorFit()}.}

\item{times}{A vector of timepoints at which to compute fitted response values}

\item{fitType}{String indicating whether to calculate statistics on the
posterior mean, posterior samples, or raw model fit.}

\item{features}{a vector of selected feature names, row numbers, or logical
values for which to calculate fitted values}
}
\value{
a data.table with columns for time and fitted value for each feature.
}
\description{
Calculates the fitted values for selected features at given times.
}
