% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{getPosteriorFit}
\alias{getPosteriorFit}
\title{perform multivariate adaptive shrinkage (mashr) on a fitted model
and return the posterior fit}
\usage{
getPosteriorFit(
  fit,
  covMethod = c("data-driven", "canonical", "both"),
  getSigResArgs = list(),
  npc = fit$nKnots,
  covEdArgs = list(),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{fit}{A fitted linear model object, as provided by \code{getModelFit}
or \code{getPosteriorFit}.}

\item{covMethod}{string indicating type of covariance matrices to compute
when fitting the \code{mash} model. Can take one of 'data-driven', 'canonical', 'or both'}

\item{getSigResArgs}{list of argument to be passed to
\code{mashr::get_significant_results()}. Used to find significant effects
from at least one condition.}

\item{npc}{number of principle components to use in
principle component analysis of fitted model}

\item{covEdArgs}{list specifying argument-value pairs
to be passed to \code{mashr::cov_ed()}.}

\item{overwrite}{TRUE or FALSE to recompute mashr fit}
}
\value{
a \code{LimoRhyde2} class object containing everything found in \code{fit}
with added elements:
\itemize{
\item \code{mashData} a data object of class \code{mash}
\item \code{mashFit} list of mash fit results
\item \code{mashCoefficients} matrix of resulting posterior model coefficients
}
}
\description{
given a linear model fit, from \code{getModelFit},
compute moderated coefficients using multivariate adaptive shrinkage (mashr).
}
