% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{getPosteriorSamples}
\alias{getPosteriorSamples}
\title{Draw samples from posterior estimate distributions from a fitted model}
\usage{
getPosteriorSamples(fit, nPosteriorSamples = 200, overwrite = FALSE)
}
\arguments{
\item{fit}{A limorhyde2 object, as provided
by \code{getModelFit()} or \code{getPosteriorFit()}.}

\item{nPosteriorSamples}{number of samples to draw
from the posterior distribution of each effect}

\item{overwrite}{Logical for whether to recompute posterior samples if it already exists.}
}
\value{
a limorhyde2 object containing
everything found in \code{fit} with added element:
\itemize{
\item \code{mashPosteriorSamples}  an m x n x p array with
m features, n model coefficients, and p posterior samples
}
}
\description{
After applying a mash model with \code{getPosteriorFit()}, which computes
posterior distribution of each effect, use \code{getPosteriorSamples} to
sample from the posterior for a set of estimates for each coefficient.
}
