% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{getRhythmStats}
\alias{getRhythmStats}
\title{Calculate rhythmic statistics from a fitted model}
\usage{
getRhythmStats(
  fit,
  fitType = c("posterior_mean", "posterior_samples", "raw"),
  features = NULL
)
}
\arguments{
\item{fit}{A limorhyde2 fit object, as provided by \code{getModelFit} or
\code{getPosteriorFit}.}

\item{fitType}{String indicating whether to calculate statistics on the
posterior mean, posterior samples, or raw model fit. Can be one of
'posterior_mean', 'posterior_samples', or 'raw'.}

\item{features}{Vector containing the names, row numbers, or logical conditions for
features on which to calculate rhythmic statistics.}
}
\value{
A data.table with rows for each feature and columns for rhythmic statistic
(mean value, peak-to-trough amplitude, root mean-squared amplitude,
peak phase, and trough phase), as well as condition and posterior sample,
if applicable.
}
\description{
\code{getRhythmStats}returns the rhythmic statistics calculated under a
fitted linear model for selected features.
}
