% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{getStatsIntervals}
\alias{getStatsIntervals}
\title{Calculate credible intervals for rhythmic statistics}
\usage{
getStatsIntervals(posteriorStats, mass = 0.9, method = c("eti", "hdi"))
}
\arguments{
\item{posteriorStats}{A data.table of posterior samples of rhythmic statistics.}

\item{mass}{The probability mass for which to calculate the interval.}

\item{method}{One of 'eti' or 'hdi'. Inputting 'eti' returns an equal-tailed
interval (i.e., an interval with an equal probability mass in each tail);
inputting 'hdi' returns a highest-posterior density interval (i.e., the
narrowest interval containing the specified probability mass).}
}
\value{
A data.table with rows for each feature and rhythmic statistic,
and columns for the upper and lower bounds of the credible interval.
}
\description{
\code{getStatsIntervals} constructs credible intervals for each of the rhythmic
statistics calculated in \code{getRhythmStats}.
}
