% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE34018}
\alias{GSE34018}
\title{Gene expression data for GSE34018}
\format{
A list with two elements:
\itemize{
\item \code{y}: Matrix of normalized, log-transformed expression values. Rows
correspond to genes (rownames are Entrez Gene IDs) and columns to samples.
\item \code{metadata}: \code{data.table} with one row per sample. \code{time} is in hours.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE34018}
}
\usage{
GSE34018
}
\description{
Data are based on total RNA, measured by microarray, obtained from livers of
wild-type and liver-specific Reverb alpha/beta double knockout mice at
various times in a 12h:12h light:dark cycle. To save space and time, the data
include only a subset of genes, and so are mainly useful for examples of how
to use \code{limorhyde2}.
}
\seealso{
\link{GSE54650}, \code{\link[=getModelFit]{getModelFit()}}
}
\keyword{datasets}
