% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE54650}
\alias{GSE54650}
\title{Gene expression data for GSE54650}
\format{
A list with two elements:
\itemize{
\item \code{y}: Matrix of normalized, log-transformed expression values. Rows
correspond to genes (rownames are Entrez Gene IDs) and columns to samples.
\item \code{metadata}: data.table with one row per sample. \code{time} is in hours.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE54650}
}
\usage{
GSE54650
}
\description{
Data are based on total RNA, measured by microarray, obtained from livers of
wild-type mice at various times after transfer to constant darkness. To save
space and time, the data include only a subset of genes, and so are mainly
useful for examples of how to use \code{limorhyde2}.
}
\seealso{
\link{GSE34018}, \code{\link[=getModelFit]{getModelFit()}}
}
\keyword{datasets}
