% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{getDiffRhythmStats}
\alias{getDiffRhythmStats}
\title{Calculate differential rhythmic statistics between conditions}
\usage{
getDiffRhythmStats(fit, rhyStats, condLevels)
}
\arguments{
\item{fit}{A fitted linear model object, as provided by \code{getModelFit} or
\code{getPosteriorFit}.}

\item{rhyStats}{A data.table of rhythmic statistics, as returned by
\code{getRhythmStats}.}

\item{condLevels}{A vector of strings listing the pair of conditions to be compared.}
}
\value{
A data.table of differential rhythmic statistics between paired conditions.
}
\description{
\code{getDiffRhythmStats} returns the differential rhythmic statistics
calculated for subjects under differing conditions.
}
