% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitted.R
\name{getFittedValues}
\alias{getFittedValues}
\title{Compute the predicted expression values from a linear fitted model}
\usage{
getFittedValues(
  fit,
  times,
  fitType = c("posterior_mean", "posterior_samples", "raw"),
  features = NULL
)
}
\arguments{
\item{fit}{A fitted linear model object, as provided by
\code{getModelFit} or \code{getPosteriorFit}.}

\item{times}{A vector of timepoints at which to compute fitted response values}

\item{fitType}{String indicating whether to calculate statistics on the
posterior mean, posterior samples, or raw model fit. Can be one of
posterior_mean', 'posterior_samples', or 'raw'.}

\item{features}{a vector of selected feature indices to calculate fitted values}
}
\value{
a data.table with time and fitted expression values columns for each feature.
}
\description{
Calculates the fitted values for each gene at given times.
If shifted models are given, average fitted values at the given timepoints.
}
