% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{getModelFit}
\alias{getModelFit}
\title{Fit a linear model to the measurements of each feature}
\usage{
getModelFit(
  y,
  metadata,
  period = 24,
  nKnots = 4,
  timeColname = "time",
  condColname = NULL,
  covarColnames = NULL,
  nShifts = 3,
  method = c("trend", "voom"),
  lmFitArgs = list(),
  eBayesArgs = if (method == "trend") list(trend = TRUE) else list()
)
}
\arguments{
\item{y}{A matrix-like data object where each row is a feature and each column
corresponds to a time sample.}

\item{metadata}{A data.table specifying experimental design information for
each sample. Each row is a sample with metadata given in columns.}

\item{period}{Number specifying the period for time variable.
Must be same unit as sample timepoints.}

\item{nKnots}{Number of knots or internal breakpoints of periodic spline.}

\item{timeColname}{String of column in \code{metadata} with the time each sample
was acquired.}

\item{condColname}{String indicating column in \code{metadata} with condition/group
name(if any) for each sample.}

\item{covarColnames}{String vector of covariate column names in \code{metadata}
to include in linear model.}

\item{nShifts}{Number of times to offset or shift time vector. Model will fit
data using each new shifted time vector.}

\item{method}{String indicating the fitting method for \code{\link[limma:lmFit]{limma::lmFit()}}.
Takes one of 'trend' or 'voom'.}

\item{lmFitArgs}{List of arguments for \code{\link[limma:lmFit]{limma::lmFit()}}.}

\item{eBayesArgs}{List of arguments for \code{\link[limma:ebayes]{limma::eBayes()}}.}
}
\value{
A limorhyde2 object with elements:
\itemize{
\item \code{metadata}: See \code{metadata} argument above.
\item \code{timeColname}: See \code{timeColname} argument above.
\item \code{condColname}: See \code{condColname} argument above.
\item \code{covarColnames}: See \code{covarColnames} argument above.
\item \code{lmFits}: List of linear model objects with fit results for all models.
\item \code{coefficients}: A matrix with features as rows.
Columns are coefficient estimates for \code{nShift} fitted models.
\item \code{shifts}: vector of shift time values
\item \code{period}: See \code{period} argument above
\item \code{condLevels}: Vector of strings indication names of conditions if available.
\item \code{nKnots}: See \code{nKnots} argument above.
\item \code{nConds}: Number of groups or conditions.
\item \code{nCovs}: Number of covariates.
}
}
\description{
Given a data matrix where each row is a feature and each column is a time sample,
\code{getModelFit} returns a list object from fitting a linear model to the
measurements of each feature To create the linear model, the function decomposes
a periodic time variable into multiple linear components based on sine and
cosine terms or periodic spline terms of the same period.
}
\seealso{
\code{\link[limma:lmFit]{limma::lmFit()}}, \code{\link[limma:ebayes]{limma::eBayes()}}
}
