% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{getPosteriorFit}
\alias{getPosteriorFit}
\title{Perform multivariate adaptive shrinkage (mash) on a fitted model}
\usage{
getPosteriorFit(
  fit,
  covMethod = c("data-driven", "canonical", "both"),
  getSigResArgs = list(),
  npc = fit$nKnots,
  covEdArgs = list(),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{fit}{A fitted linear model object, as provided by \code{getModelFit()}
or \code{getPosteriorFit()}.}

\item{covMethod}{string indicating types of covariance matrices to compute
when fitting the \code{mash} model. Can take one of 'data-driven', 'canonical', or 'both'}

\item{getSigResArgs}{list of argument to be passed to
\link[mashr:get_significant_results]{mashr::get_significant_results}. Used to find significant effects
from at least one condition.}

\item{npc}{number of principle components to use in
principle component analysis of fitted model}

\item{covEdArgs}{list specifying argument-value pairs
to be passed to \link[mashr:cov_ed]{mashr::cov_ed}.}

\item{overwrite}{Logical for whether to recompute mashr fit if it already exists.}
}
\value{
a \code{LimoRhyde2} class object containing everything found in \code{fit}
with added elements:
\itemize{
\item \code{mashData} a data object of class \code{mash}
\item \code{mashFit} list of mash fit results
\item \code{mashCoefficients} matrix of resulting posterior model coefficients
\item \code{mashIdx} vector of fit coefficient indices included in \code{mash}fitting
}
}
\description{
Given a linear model fit, from \code{getModelFit()},
\code{getPosteriorFit()} fits a mash model to coefficients to improve coefficient estimates.
The mash model returns moderated coefficients as Posterior Means.
}
\seealso{
\code{\link[=getModelFit]{getModelFit()}}, \code{\link[mashr:get_significant_results]{mashr::get_significant_results()}},
\code{\link[mashr:cov_ed]{mashr::cov_ed()}}
}
