% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{getPosteriorSamples}
\alias{getPosteriorSamples}
\title{Draw samples from posterior estimate distributions from a fitted model}
\usage{
getPosteriorSamples(fit, nPosteriorSamples = 200, overwrite = FALSE)
}
\arguments{
\item{fit}{A fitted linear model object, as provided
by \code{getModelFit} or \code{getPosteriorFit}.}

\item{nPosteriorSamples}{number of samples to draw
from the posterior distribution of each effect}

\item{overwrite}{TRUE or FALSE to replace existing posteriorSamples}
}
\value{
a \code{LimoRhyde2} class object containing
everything found in \code{fit} with added element:
\itemize{
\item \code{mashPosteriorSamples}  an m x n x p array with
m features, n model coefficients, and p posterior samples
}
}
\description{
After performing shrinkage on a linear model with \code{getPosteriorFit},
draw samples from the posterior distribution of estimates.
}
