% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{getRhythmStats}
\alias{getRhythmStats}
\title{Calculate rhythmic statistics from a fitted model}
\usage{
getRhythmStats(
  fit,
  fitType = c("posterior_mean", "posterior_samples", "raw"),
  features = NULL
)
}
\arguments{
\item{fit}{A fitted linear model object, as provided by \code{getModelFit} or
\code{getPosteriorFit}.}

\item{fitType}{String indicating the type of linear model. Can be one of
'posterior_mean', 'posterior_samples', or 'raw'.}

\item{features}{Vector containing the names of the features for which to
calculate rhythmic statistics.}
}
\value{
A data.table with columns for each feature, rhythmic statistic, as well
as condition and posterior sample, if applicable.
}
\description{
\code{getRhythmStats} returns the rhythmic statistics (i.e., peak-to-trough
amplitude, root mean-squared amplitude, peak phase, and trough phase)
calculated under a fitted linear model for selected features.
}
