% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{getStatsIntervals}
\alias{getStatsIntervals}
\title{Calculate credible intervals for rhythmic statistics}
\usage{
getStatsIntervals(posteriorStats, mass = 0.9, method = c("eti", "hdi"))
}
\arguments{
\item{posteriorStats}{A data.table of posterior samples of (differential)
rhythmic statistics from \code{getRhythmStats} or \code{getDiffRhythmStats}.}

\item{mass}{The probability mass for which to calculate the interval.}

\item{method}{One of 'eti' or 'hdi'. Inputting 'eti' returns an equal-tailed
interval and inputting 'hdi' returns a highest-posterior density interval.}
}
\value{
A data.table with rows for each feature and (differential)
rhythmic statistic, as well as condition if applicable. Columns correspond to
the upper and lower bounds of the credible interval.
}
\description{
\code{getStatsIntervals} constructs credible intervals for each of the rhythmic
statistics calculated in \code{getRhythmStats} or differential rhythmic statistics
calculated in \code{getDiffRhythmStats}.
}
\seealso{
\code{\link[=getRhythmStats]{getRhythmStats()}}, \link{getDiffRhythmStats},
\code{\link[stats]{quantiles}} ,\code{\link[HDInterval]{hdi}}
}
